ai-receptionist/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
jinja2
python-dateutil
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
twilio
